Improved bytecode machine on int slice instead of bytes

+++ vm/opcodes.go
@@ -1,7 +1,7 @@
 package vm
 
 const (
-	OpPush byte = iota
+	OpPush int = iota
 	OpPop
 	OpRot
 	OpFetch
